move getCookie to helpers
