+++ fec/fec/static/js/modules/feedback.js
@@ -4,6 +4,7 @@ var $ = require('jquery');
 var _ = require('underscore');
 
 var accessibility = require('./accessibility');
+var helpers = require('./helpers');
 
 var feedback = require('../templates/feedback.hbs');
 
@@ -11,36 +12,6 @@ var statusClasses = {
   success: 'message--success',
   error: 'message--error'
 };
-
-/**
- * setup JQuery's AJAX methods to setup CSRF token in the request before sending it off.
- * http://stackoverflow.com/questions/5100539/django-csrf-check-failing-with-an-ajax-post-request
- */
-function getCookie(name) {
-    var cookieValue = null;
-    if (document.cookie && document.cookie != '') {
-        var cookies = document.cookie.split(';');
-        for (var i = 0; i < cookies.length; i++) {
-            var cookie = jQuery.trim(cookies[i]);
-            // Does this cookie string begin with the name we want?
-
-            if (cookie.substring(0, name.length + 1) == (name + '=')) {
-                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                break;
-            }
-        }
-    }
-    return cookieValue;
-}
-
-$.ajaxSetup({
-     beforeSend: function(xhr, settings) {
-         if (!(/^http:.*/.test(settings.url) || /^https:.*/.test(settings.url))) {
-             // Only send the token to relative URLs i.e. locally.
-             xhr.setRequestHeader("X-CSRFToken", getCookie('csrftoken'));
-         }
-     }
-});
 
 /**
  * Feedback widget
@@ -93,7 +64,21 @@ Feedback.prototype.hide = function() {
 };
 
 Feedback.prototype.submit = function(e) {
+  /**
+   * setup JQuery's AJAX methods to setup CSRF token in the request before sending it off.
+   * http://stackoverflow.com/questions/5100539/django-csrf-check-failing-with-an-ajax-post-request
+   */
+  $.ajaxSetup({
+     beforeSend: function(xhr, settings) {
+       if (!(/^http:.*/.test(settings.url) || /^https:.*/.test(settings.url))) {
+           // Only send the token to relative URLs i.e. locally.
+           xhr.setRequestHeader('X-CSRFToken', helpers.getCookie('csrftoken'));
+       }
+     }
+  });
+
   e.preventDefault();
+
   var data = _.chain(this.$box.find('textarea'))
     .map(function(elm) {
       var $elm = $(elm);
@@ -101,6 +86,7 @@ Feedback.prototype.submit = function(e) {
     })
     .object()
     .value();
+
   if (!_.some(_.values(data))) {
     var message =
       '<h2 class="feedback__title">Input required</h2>' +
@@ -117,6 +103,7 @@ Feedback.prototype.submit = function(e) {
     contentType: 'application/json',
     dataType: 'json'
   });
+
   promise.done(this.handleSuccess.bind(this));
   promise.fail(this.handleError.bind(this));
 };
