+++ fec/fec/static/js/modules/helpers.js
@@ -416,6 +416,22 @@ function sanitizeQueryParams(query) {
   return query;
 }
 
+function getCookie(name) {
+  var cookieValue = null;
+  if (document.cookie && document.cookie != '') {
+    var cookies = document.cookie.split(';');
+    for (var i = 0; i < cookies.length; i++) {
+      var cookie = $.trim(cookies[i]);
+      // Does this cookie string begin with the name we want?
+      if (cookie.substring(0, name.length + 1) == (name + '=')) {
+        cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+        break;
+      }
+    }
+  }
+  return cookieValue;
+}
+
 module.exports = {
   buildAppUrl: buildAppUrl,
   buildUrl: buildUrl,
@@ -441,5 +457,6 @@ module.exports = {
   BREAKPOINTS: BREAKPOINTS,
   getWindowWidth: getWindowWidth,
   sanitizeValue: sanitizeValue,
-  sanitizeQueryParams: sanitizeQueryParams
+  sanitizeQueryParams: sanitizeQueryParams,
+  getCookie: getCookie
 };
